feat: move DumpFixer from root to src directory
#14
