+++ composer.json
@@ -26,7 +26,7 @@
     },
     "autoload": {
         "psr-4": {
-            "whatwedo\\PhpCodingStandard\\": ""
+            "whatwedo\\PhpCodingStandard\\":  "src/"
         }
     },
     "config": {
