+++ src/Fixer/DumpFixer.php
@@ -27,7 +27,7 @@ declare(strict_types=1);
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-namespace whatwedo\PhpCodingStandard;
+namespace whatwedo\PhpCodingStandard\Fixer;
 
 use PhpCsFixer\AbstractFunctionReferenceFixer;
 use PhpCsFixer\FixerDefinition\CodeSample;
